requirements-test.txt
django>=1.8.0
coverage
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0

django-cms>=3.0
